pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bikescout"
version = "0.6.0"
authors = [
    { name="Bikescout", email="bikescout.mcp@gmail.com" },
]
description = "An MCP server for smart cycling route discovery."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]",
    "requests",
    "python-dotenv"
]

[project.scripts]
bikescout = "bikescout.mcp_server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/bikescout"]